Modify everything to use time::Now() instead of millis() to allow constant time signature across a single loop

+++ include/games/head2head/dot.h
@@ -22,7 +22,7 @@ class H2HDot : public animation::Animation {
 
   void Move() override { physics.Move(); }
 
-  void setVelocity(float velocity) { physics.velocity.y = (int)velocity; }
+  void setVelocity(float velocity) { physics.velocity.y = velocity; }
 
   void draw(display::Display* display) {
     // Don't draw outside the display boundaries
